news-ai-service/requirements.txt
flask
transformers
torch
python-dotenv
requests